% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterDendro.R
\name{clusterDendro}
\alias{clusterDendro}
\title{Plot dendrogram of clustered gene sets}
\usage{
clusterDendro(cluster_result = NULL)
}
\arguments{
\item{cluster_result}{output from clusterSets()}
}
\value{
a ggplot object(for one-part enrichments like hypergeometric) or a list of ggplot objects (for two-part like GSEA)
}
\description{
Plot dendrogram of clustered gene sets
}
\examples{
# Not run
# Create res object as described in clusterSets example
# clusterDendro(res)
}
