% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterScatter.R
\name{clusterScatter}
\alias{clusterScatter}
\title{Create scatterplot of clustered gene set result}
\usage{
clusterScatter(cluster_result = NULL, dimred = "UMAP", scores = "pvalue")
}
\arguments{
\item{cluster_result}{the result of clusterSets which contains the overlap coefficient distance matrix and the cluster membership of gene sets to be plotted}

\item{dimred}{dimension reduction method. Options are "UMAP", "tSNE", or "PCoA".}

\item{scores}{How should points be sized on the scatterplot? Options are "pvalue" and "size". Pvalue will scale points according to -log10(pvalue), size will scale points according to gene set size K}
}
\value{
a ggplot object(for one-part enrichments like hypergeometric) or a list of ggplot objects (for two-part like GSEA)
}
\description{
Create scatterplot of clustered gene set result
}
\examples{
# Not run
# Create res object as described in clusterSets example
# clusterScatter(res, dimred = "UMAP", scores = "size")
}
