% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTreemap.R
\name{clusterTreemap}
\alias{clusterTreemap}
\title{Create treemap plot(s) of clustered enrichment result}
\usage{
clusterTreemap(cluster_result = NULL, scores = "pvalue")
}
\arguments{
\item{cluster_result}{the result of clusterSets which contains the overlap coefficient distance matrix and the cluster membership of gene sets to be plotted}

\item{scores}{How should points be sized on the scatterplot? Options are "pvalue" and "size". Pvalue will scale points according to -log10(pvalue), size will scale points according to gene set size K}
}
\value{
a ggplot object(for one-part enrichments like hypergeometric) or a list of ggplot objects (for two-part like GSEA)
}
\description{
Create treemap plot(s) of clustered enrichment result
}
\examples{
# Not run
# Create res object as described in clusterSets example
# clusterTreemap(res)
}
