% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat.rda}
\alias{dat.rda}
\alias{dat}
\title{madRich example data}
\format{
A data frame with 100 rows and 8 variables:
\describe{
\item{group}{character. Value "g1" or "g2"}
\item{pathway}{character. Pathway names from msigdb}
\item{gs_cat}{character. Gene set category from msigdb. Values "H", "C2", or"C5"}
\item{gs_subcat}{character. Gene set subcategory from msigdb. Values NA, "CP", or"GO:BP"}
\item{NES}{double. Normalized enrichment score for GSEA-type analysis}
\item{k/K}{double. k/K ratio for hypergeometric-type analysis}
\item{pvalue}{double. P-value for enrichment test}
\item{FDR}{double. Corrected p-value for enrichment test}
}
}
\usage{
dat
}
\description{
A small example data frame
}
\keyword{datasets}
