% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterDendro.R
\name{clusterDendro}
\alias{clusterDendro}
\title{Plot dendrogram of clustered gene sets}
\usage{
clusterDendro(cluster_result = NULL)
}
\arguments{
\item{cluster_result}{output from clusterSets()}
}
\value{
a ggplot object(for one-part enrichments like hypergeometric) or a list of ggplot objects (for two-part like GSEA)
}
\description{
Plot dendrogram of clustered gene sets
}
\examples{
#' # Run enrichment using SEARchways
gene_list2 <- list(HRV1 = names(SEARchways::example.gene.list[[1]]),
                  HRV2 = names(SEARchways::example.gene.list[[2]]))
df1 <- SEARchways::BIGprofiler(gene_list=gene_list2, 
                             category="C5", subcategory="GO:MF", ID="ENSEMBL")
df2 <- SEARchways::BIGprofiler(gene_list=gene_list2, 
                              category="C5", subcategory="GO:BP", ID="ENSEMBL")
df <- dplyr::bind_rows(df1, df2)
res <- clusterSets(df = df, enrich_method="hypergeometric",
                   ID = "ENSEMBL",
                   collections = c("C5"),
                   subcollections = c("C5" = "GO:MF", "C5" = "GO:BP"),
                   hclust_height = c(0.7),
                   group_name = "HRV1",
                   fdr_cutoff = 0.4)
clusterDendro(res)
}
