% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterDendro.R
\name{clusterDendro}
\alias{clusterDendro}
\title{Plot dendrogram of clustered gene sets}
\usage{
clusterDendro(cluster_result = NULL)
}
\arguments{
\item{cluster_result}{output from clusterSets()}
}
\value{
a ggplot object(for one-part enrichments like hypergeometric) or a list of ggplot objects (for two-part like GSEA)
}
\description{
Plot dendrogram of clustered gene sets
}
\examples{
res <- clusterSets(df = dat,
                   collections = c("H", "C2", "C5"),
                   subcollections = c("C2" = "CP", "C5" = "GO:BP"),
                   hclust_height = 0.9,
                   enrich_method = "hypergeometric")
clusterDendro(res)
}
