% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterScatter.R
\name{clusterScatter}
\alias{clusterScatter}
\title{Create scatterplot of clustered gene set result}
\usage{
clusterScatter(cluster_result = NULL, dimred = "UMAP", scores = "pvalue")
}
\arguments{
\item{cluster_result}{the result of clusterSets which contains the overlap coefficient distance matrix and the cluster membership of gene sets to be plotted}

\item{dimred}{dimension reduction method. Options are "UMAP", "tSNE", or "PCoA".}

\item{scores}{How should points be sized on the scatterplot? Options are "pvalue" and "size". Pvalue will scale points according to -log10(pvalue), size will scale points according to gene set size K}
}
\value{
a ggplot object(for one-part enrichments like hypergeometric) or a list of ggplot objects (for two-part like GSEA)
}
\description{
Create scatterplot of clustered gene set result
}
\examples{
#' # Run enrichment using SEARchways
gene_list2 <- list(HRV1 = names(SEARchways::example.gene.list[[1]]),
                  HRV2 = names(SEARchways::example.gene.list[[2]]))
df1 <- SEARchways::BIGprofiler(gene_list=gene_list2, 
                             category="C5", subcategory="GO:MF", ID="ENSEMBL")
df2 <- SEARchways::BIGprofiler(gene_list=gene_list2, 
                              category="C5", subcategory="GO:BP", ID="ENSEMBL")
df <- dplyr::bind_rows(df1, df2)
res <- clusterSets(df = df, enrich_method="hypergeometric",
                   ID = "ENSEMBL",
                   collections = c("C5"),
                   subcollections = c("C5" = "GO:MF", "C5" = "GO:BP"),
                   hclust_height = c(0.7),
                   group_name = "HRV1",
                   fdr_cutoff = 0.4)
clusterScatter(res, dimred = "PCoA", scores = "size")
}
