% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSets.R
\name{clusterSets}
\alias{clusterSets}
\title{Perform Hierarchical Clustering of Gene Sets Based on the Overlap Coefficient}
\usage{
clusterSets(
  df = NULL,
  enrich_method = "hypergeometric",
  collections = NULL,
  subcollections = NULL,
  db = NULL,
  ID = "SYMBOL",
  species = "HS",
  hclust_height = 0.7,
  fdr_cutoff = NULL,
  abs_NES_cutoff = NULL,
  kK_cutoff = NULL,
  group_name = NULL
)
}
\arguments{
\item{df}{result of enrichment function. Should contain a column of pathway names ("pathway"), p-values ("pvalue"), gene set collections ("gs_collection"), and gene set subcollections where applicable ("gs_subcollection"). If filtering by FDR, k/K, or NES, these values must be included as "FDR", "k/K", or "NES", respectively.}

\item{enrich_method}{"hypergeometric" or "gsea." "gsea" will separate result by positive/negative NES within your filtered df.}

\item{collections}{a vector of Broad gene set categories included among the pathway names in df}

\item{subcollections}{a named vector of Broad gene set subcategories (names are corresponding cats provided in 'collections')}

\item{db}{optional, custom gene set database formatted like MSigDB}

\item{ID}{"SYMBOL", "ENSEMBL", or "ENTREZ". What format of gene ID do you want to use for clustering.}

\item{species}{"HS" for human, or "MM" for mouse for msigDB}

\item{hclust_height}{Height for cutting tree for hclust. Value must be between 0 and 1}

\item{fdr_cutoff}{optional, high-end filter to apply to df before clustering}

\item{abs_NES_cutoff}{optional, low-end. absolute value filter to apply to df before clustering of GSEA results}

\item{kK_cutoff}{optional, low-end filter to apply to df before clustering of hypergeometric results}

\item{group_name}{optional, if your dataset contains multiple groups in a 'group' column, select pathways from provided group. Otherwise, all pathways passing filter will be clustered.}
}
\value{
a list containing the input data frame, formatted reference database, cluster membership, and distance matrix/matrices. To be used in downstream visualizations.
}
\description{
Perform Hierarchical Clustering of Gene Sets Based on the Overlap Coefficient
}
\examples{
res <- clusterSets(df = dat,
                   collections = c("H", "C2", "C5"),
                   subcollections = c("C2" = "CP", "C5" = "GO:BP"),
                   hclust_height = 0.5,
                   enrich_method = "gsea",
                   group_name = "g1")
}
