% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTreemap.R
\name{clusterTreemap}
\alias{clusterTreemap}
\title{Create treemap plot(s) of clustered enrichment result}
\usage{
clusterTreemap(cluster_result = NULL, scores = "pvalue")
}
\arguments{
\item{cluster_result}{the result of clusterSets which contains the overlap coefficient distance matrix and the cluster membership of gene sets to be plotted}

\item{scores}{How should points be sized on the scatterplot? Options are "pvalue" and "size". Pvalue will scale points according to -log10(pvalue), size will scale points according to gene set size K}
}
\value{
a ggplot object(for one-part enrichments like hypergeometric) or a list of ggplot objects (for two-part like GSEA)
}
\description{
Create treemap plot(s) of clustered enrichment result
}
\examples{
res <- clusterSets(df = dat,
                   category = c("H", "C2", "C5"),
                   subcategory = c("C2" = "CP", "C5" = "GO:BP"),
                   hclust_height = 0.9,
                   enrich_method = "hypergeometric")
clusterTreemap(res)
}
