% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterWordclouds.R
\name{clusterWordclouds}
\alias{clusterWordclouds}
\title{Create word cloud plots to help in annotation of gene set clusters}
\usage{
clusterWordclouds(cluster_result = NULL, rmwords = NULL)
}
\arguments{
\item{cluster_result}{the result of clusterSets which contains the overlap coefficient distance matrix and the cluster membership of gene sets to be plotted}

\item{rmwords}{a vector of words to remove from wordclouds. a standard list of words is included (articles, common broad category names, etc), but if your results have unexpected emphasis of useless words, you can add more here.}
}
\value{
a list of ggplot objects
}
\description{
Create word cloud plots to help in annotation of gene set clusters
}
\examples{
#' # Run enrichment using SEARchways
gene_list2 <- list(HRV1 = names(SEARchways::example.gene.list[[1]]),
                  HRV2 = names(SEARchways::example.gene.list[[2]]))
df1 <- SEARchways::BIGprofiler(gene_list=gene_list2, 
                             category="C5", subcategory="GO:MF", ID="ENSEMBL")
df2 <- SEARchways::BIGprofiler(gene_list=gene_list2, 
                              category="C5", subcategory="GO:BP", ID="ENSEMBL")
df <- dplyr::bind_rows(df1, df2)
res <- clusterSets(df = df, enrich_method="hypergeometric",
                   ID = "ENSEMBL",
                   collections = c("C5"),
                   subcollections = c("C5" = "GO:MF", "C5" = "GO:BP"),
                   hclust_height = c(0.7),
                   group_name = "HRV1",
                   fdr_cutoff = 0.4)
clusterWordclouds(res, rmwords = c("defense", "immune"))
}
