% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterWordclouds.R
\name{clusterWordclouds}
\alias{clusterWordclouds}
\title{Create word cloud plots to help in annotation of gene set clusters}
\usage{
clusterWordclouds(cluster_result = NULL, rmwords = NULL)
}
\arguments{
\item{cluster_result}{the result of clusterSets which contains the overlap coefficient distance matrix and the cluster membership of gene sets to be plotted}

\item{rmwords}{a vector of words to remove from wordclouds. a standard list of words is included (articles, common broad category names, etc), but if your results have unexpected emphasis of useless words, you can add more here.}
}
\value{
a list of ggplot objects
}
\description{
Create word cloud plots to help in annotation of gene set clusters
}
\examples{
# Not run
# Create res object as described in clusterSets example
# clusterWordclouds(res, rmwords = c("defense", "immune"))
}
